Move TransactionStatus to TransactionStorage.kt
